requirements.txt
absl-py==2.1.0
accelerate==1.1.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
datasets==3.1.0
decorator==4.4.2
Deprecated==1.2.15
diffusers==0.27.2
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
fastjsonschema==2.21.0
ffmpy==0.4.0
filelock==3.13.1
fire==0.7.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.6.0
gradio_client==1.4.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
icecream==2.1.3
idna==3.4
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
iopath==0.1.10
ipycanvas==0.13.3
ipyevents==2.0.2
ipython==8.30.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jaxtyping==0.2.34
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
ml_dtypes==0.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
#nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@729261dc64c4241ea36efda84fbf532cc8b425b8
#nvidia-cuda-runtime-cu12==12.6.77
nvidia-pyindex==1.0.9
#nvidia-tensorrt==99.0.0
omegaconf==2.3.0
onnxruntime==1.20.0
onnxruntime-gpu==1.20.0
open3d==0.18.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.11
#ort-nightly-gpu==1.15.0.dev20230502003
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.13.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
portalocker==2.10.1
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pygltflib==1.16.2
Pygments==2.18.0
PyMatting==1.1.13
#pymeshlab==2023.12.post2
pyparsing==3.2.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
#pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@75ebeeaea0908c5527e7b1e305fbc7681382db47
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
rembg==2.0.59
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.21.0
ruff==0.7.4
safehttpx==0.1.1
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.18.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
spaces==0.30.4
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.1
tenacity==9.0.0
# tensorrt==10.6.0
# tensorrt-cu12==10.6.0
# tensorrt-cu12-bindings==10.6.0
# tensorrt-cu12-libs==10.6.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.3
tomlkit==0.12.0
#torch_scatter==2.1.2
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
trimesh==4.5.2
triton==2.1.0
typeguard==2.13.3
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.2
urllib3==1.26.13
usd-core==24.11
uvicorn==0.32.0
wandb==0.18.7
warp-lang==1.4.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.16.0
xformers==0.0.22.post7
xxhash==3.5.0
yacs==0.1.8
yarl==1.17.1
zipp==3.21.0

# The user requested huggingface-hub==0.24.5
# accelerate 1.1.1 depends on huggingface-hub>=0.21.0
# datasets 3.1.0 depends on huggingface-hub>=0.23.0
# diffusers 0.27.2 depends on huggingface-hub>=0.20.2
# gradio 5.6.0 depends on huggingface-hub>=0.25.1